Split storyboard parser into files for each type
